video-processing-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
redis>=5.0.0
google-generativeai>=0.8.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pillow>=10.0.0
yt-dlp>=2024.7.1
rich>=13.0.0
tiktoken>=0.5.0
python-multipart>=0.0.6
aiofiles>=23.0.0
httpx>=0.26.0
aiohttp>=3.8.0
cachetools>=5.3.0
numpy>=1.24.0